build: Adds last frame thumbnail for images
