+++ iosApp/iosApp/Feed/View/FeedsVC/FeedsViewController+ScrollView.swift
@@ -1,0 +1,57 @@
+//
+//  FeedsViewController+ScrollView.swift
+//  iosApp
+//
+//  Created by Sarvesh Sharma on 20/01/25.
+//  Copyright © 2025 orgName. All rights reserved.
+//
+import UIKit
+
+extension FeedsViewController: UICollectionViewDelegate {
+  func scrollViewWillBeginDragging(_ scrollView: UIScrollView) {
+    guard scrollView.contentOffset.y > .zero else {
+      storeThumbnail()
+      return
+    }
+    yralPlayer.pause()
+    storeThumbnail()
+  }
+
+  func scrollViewDidEndDecelerating(_ scrollView: UIScrollView) {
+    feedsCV.layoutIfNeeded()
+    guard let visibleIndexPath = feedsCV.indexPathsForVisibleItems.sorted().first else { return }
+    let oldIndex = yralPlayer.currentIndex
+    let newIndex = visibleIndexPath.item
+    if newIndex != oldIndex {
+      yralPlayer.advanceToVideo(at: newIndex)
+      feedsCV.reloadData()
+    } else {
+      yralPlayer.play()
+    }
+  }
+
+  func scrollViewShouldScrollToTop(_ scrollView: UIScrollView) -> Bool {
+    return false
+  }
+
+  func collectionView(
+    _ collectionView: UICollectionView,
+    willDisplay cell: UICollectionViewCell,
+    forItemAt indexPath: IndexPath
+  ) {
+    let feedsCount = feedsDataSource.snapshot().numberOfItems
+    if indexPath.item >= feedsCount - Constants.thresholdForLoadingMoreResults, !loadMoreRequestMade {
+      Task {
+        self.loadMoreRequestMade = true
+        await viewModel.loadMoreFeeds()
+      }
+    }
+    if indexPath.item >= feedsCount - 1 {
+      DispatchQueue.main.async {
+        self.shouldShowFooterLoader = true
+        let snapshot = self.feedsDataSource.snapshot()
+        self.feedsDataSource.apply(snapshot, animatingDifferences: false)
+      }
+    }
+  }
+}
